+++ dev/Network.js
@@ -2,7 +2,7 @@
 
 class Network {
 
-    constructor ({learningRate, layers=[], adaptiveLR="noAdaptiveLR", activation="sigmoid", cost="crossEntropy", rmsDecay, rho}={}) {
+    constructor ({learningRate, layers=[], adaptiveLR="noAdaptiveLR", activation="sigmoid", cost="crossEntropy", rmsDecay, rho, lreluSlope}={}) {
         this.state = "not-defined"
         this.layers = []
         this.epochs = 0
@@ -31,6 +31,7 @@ class Network {
                 if(this.learningRate==undefined){
                     switch(activation) {
                         case "relu":
+                        case "lrelu":
                             this.learningRate = 0.01
                             break
                         case "tanh":
@@ -44,11 +45,15 @@ class Network {
         
         this.adaptiveLR = [false, null, undefined].includes(adaptiveLR) ? "noAdaptiveLR" : adaptiveLR
         this.weightUpdateFn = NetMath[this.adaptiveLR]
-        this.activation = NetMath[activation]
+        this.activation = NetMath[activation].bind(this)
         this.cost = NetMath[cost]
 
         if(this.adaptiveLR=="RMSProp"){
             this.rmsDecay = rmsDecay==undefined ? 0.99 : rmsDecay
+        }
+
+        if(activation=="lrelu"){
+            this.lreluSlope = lreluSlope==undefined ? -0.0005 : lreluSlope
         }
 
         if(layers.length) {