+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 #### NetMath
 - Added tanh activation function
 - Added relu activation function
+- Added lrelu activation function
+- Added lreluSlope as a network configuration
 # v1.1.0 - Adaptive learning rates
 ---
 #### Network